requirements.txt
# AI Employee FTE - Python Dependencies
# Bronze Tier Requirements

# Core dependencies (required for all watchers)
watchdog>=3.0.0

# Gmail Watcher dependencies (optional - install if using Gmail integration)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# WhatsApp Watcher dependencies (optional - install if using WhatsApp integration)
# Note: Playwright requires additional browser installation
# Run: playwright install chromium
playwright>=1.40.0

# Qwen Code API Integration (required for AI automation)
openai>=1.0.0
python-dotenv>=1.0.0

# Utilities
python-dotenv>=1.0.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
